+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.3] - 2025-03-03
 ### Added
 - `TfidfVectorizer` feature extractor
 - `TfidfTransformer` feature extractor
@@ -388,7 +388,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `Export` class
 - `Import` class
 
-[Unreleased]: https://github.com/openscilab/pymilo/compare/v1.2...dev
+[Unreleased]: https://github.com/openscilab/pymilo/compare/v1.3...dev
+[1.3]: https://github.com/openscilab/pymilo/compare/v1.2...v1.3
 [1.2]: https://github.com/openscilab/pymilo/compare/v1.1...v1.2
 [1.1]: https://github.com/openscilab/pymilo/compare/v1.0...v1.1
 [1.0]: https://github.com/openscilab/pymilo/compare/v0.9...v1.0